Address feedback part 3 - rename component
